# Clientory App/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /businesses/suggest-competitors:
    post:
      operationId: suggestCompetitors
      tags: [businesses]
      summary: Suggest likely competitors using AI
      description: Uses Claude Haiku to suggest 3-5 likely competitors. No auth required. IP rate limited.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SuggestCompetitorsBody"
      responses:
        "200":
          description: Competitor suggestions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuggestCompetitorsResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RateLimitResponse"
  /businesses:
    post:
      operationId: createBusiness
      tags: [businesses]
      summary: Create or update a firm profile
      parameters:
        - name: x-email-token
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBusinessBody"
      responses:
        "201":
          description: Business created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
        "200":
          description: Business updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      operationId: listBusinesses
      tags: [businesses]
      summary: List user businesses with subscription state
      parameters:
        - name: x-email-token
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of businesses
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessListResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    RateLimitResponse:
      type: object
      properties:
        error:
          type: string
        retryAfter:
          type: number
      required:
        - error
...
    FirmLocation:
      type: object
      properties:
        city:
          type: string
        state:
          type: string
        neighborhood:
          type: string
          nullable: true
        isHQ:
          type: boolean
      required:
        - city
        - state
        - isHQ
    FirmPartner:
      type: object
      properties:
        name:
          type: string
        title:
          type: string
        trackIndependently:
          type: boolean
      required:
        - name
        - title
        - trackIndependently
    FirmCompetitor:
      type: object
      properties:
        name:
          type: string
        location:
          type: string
          nullable: true
      required:
        - name
    CreateBusinessBody:
      type: object
      properties:
        name:
          type: string
        website:
          type: string
          nullable: true
        businessType:
          type: string
        location:
          type: string
        description:
          type: string
          nullable: true
        legalName:
          type: string
          nullable: true
        brandName:
          type: string
          nullable: true
        firmType:
          type: string
          nullable: true
        yearFounded:
          type: number
          nullable: true
        partners:
          type: array
          items:
            $ref: "#/components/schemas/FirmPartner"
          nullable: true
        locations:
          type: array
          items:
            $ref: "#/components/schemas/FirmLocation"
          nullable: true
        geographicScope:
          type: string
          nullable: true
        primaryServices:
          type: array
          items:
            type: string
          nullable: true
        deliverables:
          type: array
          items:
            type: string
          nullable: true
        specialties:
          type: array
          items:
            type: string
          nullable: true
        industriesServed:
          type: array
          items:
            type: string
          nullable: true
        clientStages:
          type: array
          items:
            type: string
          nullable: true
        decisionMakers:
          type: array
          items:
            type: string
          nullable: true
        directCompetitors:
          type: array
          items:
            $ref: "#/components/schemas/FirmCompetitor"
          nullable: true
        rankings:
          type: array
          items:
            type: object
            properties:
              source:
                type: string
              category:
                type: string
              year:
                type: number
            required: [source, category, year]
          nullable: true
        topGSCQueries:
          type: array
          items:
            type: string
          nullable: true
        clientType:
          type: string
          nullable: true
        individualServices:
          type: array
          items:
            type: string
          nullable: true
        individualDeliverables:
          type: array
          items:
            type: string
          nullable: true
        individualSpecialties:
          type: array
          items:
            type: string
          nullable: true
        businessServices:
          type: array
          items:
            type: string
          nullable: true
        businessDeliverables:
          type: array
          items:
            type: string
          nullable: true
        businessSpecialties:
          type: array
          items:
            type: string
          nullable: true
    BusinessSubscriptionInfo:
      type: object
      properties:
        status:
          type: string
          enum: [active, past_due, canceled, incomplete]
        currentPeriodEnd:
          type: string
          format: date-time
        cancelAtPeriodEnd:
          type: boolean
      required:
        - status
        - currentPeriodEnd
        - cancelAtPeriodEnd
    BusinessUsageInfo:
      type: object
      properties:
        scansUsed:
          type: number
        scansLimit:
          type: number
        periodEnd:
          type: string
          format: date-time
      required:
        - scansUsed
        - scansLimit
        - periodEnd
    Business:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        website:
          type: string
          nullable: true
        businessType:
          type: string
        location:
          type: string
        description:
          type: string
          nullable: true
        legalName:
          type: string
          nullable: true
        brandName:
          type: string
          nullable: true
        firmType:
          type: string
          nullable: true
        yearFounded:
          type: number
          nullable: true
        locations:
          nullable: true
        primaryServices:
          nullable: true
        deliverables:
          nullable: true
        specialties:
          nullable: true
        industriesServed:
          nullable: true
        clientStages:
          nullable: true
        decisionMakers:
          nullable: true
        directCompetitors:
          nullable: true
        clientType:
          type: string
          nullable: true
        individualServices:
          nullable: true
        individualDeliverables:
          nullable: true
        individualSpecialties:
          nullable: true
        businessServices:
          nullable: true
        businessDeliverables:
          nullable: true
        businessSpecialties:
          nullable: true
        profileUpdatedAt:
          type: string
          format: date-time
          nullable: true
        hasPromptSet:
          type: boolean
        createdAt:
          type: string
          format: date-time
        subscription:
          $ref: "#/components/schemas/BusinessSubscriptionInfo"
          nullable: true
        usage:
          $ref: "#/components/schemas/BusinessUsageInfo"
          nullable: true
      required:
        - id
        - name
        - businessType
        - location
        - createdAt
        - hasPromptSet
    BusinessListResponse:
      type: object
      properties:
        businesses:
          type: array
          items:
            $ref: "#/components/schemas/Business"
      required:
        - businesses
...
    SuggestCompetitorsBody:
      type: object
      properties:
        firmType:
          type: string
        city:
          type: string
        state:
          type: string
        primaryServices:
          type: array
          items:
            type: string
      required:
        - firmType
        - city
        - state
    SuggestCompetitorsResponse:
      type: object
      properties:
        suggestions:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              location:
                type: string
                nullable: true
            required:
              - name
      required:
        - suggestions